2 bytes for the price of 1!
- Fix a game crash if you exploded yourself with a hydra
- Halved the number of bytes sent for md5's

TODO: linux compile

+++ ACv2_Server/CAntiCheat.cpp
@@ -395,7 +395,7 @@ void CAntiCheat::ToggleLiteFoot(bool toggle)
 	m_LiteFoot = toggle;
 }
 
-void CAntiCheat::ToggleCrouchBug(int toggle)
+void CAntiCheat::ToggleCrouchBug(unsigned short toggle)
 {
 	// Prepare to send RPC to client.
 	RakNet::BitStream bsData;
